flask/requirements.txt
flask
flask-sqlalchemy
telethon
selenium
webdriver-manager
ton-sdk
python-dotenv
requests